replace default avatar with logo component
