+++ src/common/components/index.ts
@@ -1,13 +1,13 @@
 export { BreadCrumb } from "./breadcrumb/index";
 export { Button } from "./button/index";
 export { CheckDropdown } from "./check-dropdown/index";
-export { DefaultAvatar } from "./default-avatar/index";
 export { Dropdown } from "./dropdown/index";
 export { Footer } from "./footer/index";
 export { Header } from "./header/index";
 export { Icon } from "./icon/index";
 export { IconButton } from "./icon-button/index";
 export { Input } from "./input/index";
+export { Logo } from "./logo/index";
 export { Modal } from "./modal/index";
 export { Pagination } from "./pagination/index";
 export { SearchBar } from "./search-bar/index";
