+++ src/common/components/logo/styles.module.scss
