+++ src/pages/companies-list-page/components/companies-list/components/filtered-companies-list/components/companies-list/components/company-list-card-desktop/index.tsx
@@ -2,8 +2,7 @@ import React from "react";
 import { Link } from "react-router-dom";
 
 import ShieldTick from "~/assets/images/shield-tick.svg?react";
-import { DefaultAvatar, StarRating } from "~/common/components/index";
-import { IMAGE_UPLOAD_URL, LOGO_NONE } from "~/common/constants/index";
+import { Logo, StarRating } from "~/common/components/index";
 import { AppRoute, RatingSize } from "~/common/enums/index";
 import { Company } from "~/common/types/index";
 
@@ -20,19 +19,11 @@ const CompanyListCardDesktop: React.FC<Properties> = ({ company }) => {
 			to={`${AppRoute.COMPANY_DETAILS}${company.id}`}
 		>
 			<div className={styles["company_list_card__title"]}>
-				{company.logo !== LOGO_NONE ? (
-					<img
-						alt={`${company.name} logo`}
-						className={styles["company_list_card__logo"]}
-						src={`${IMAGE_UPLOAD_URL}${company.logo}`}
-					/>
-				) : (
-					<DefaultAvatar
-						className={styles["company_list_card__logo"]}
-						name={company.name}
-					/>
-				)}
-
+				<Logo
+					className={styles["company_list_card__logo"]}
+					logo={company.logo}
+					name={company.name}
+				/>
 				<span className={styles["company_list_card__name"]}>
 					{company.name}
 				</span>
