+++ src/pages/companies-list-page/components/companies-list/components/filtered-companies-list/components/companies-list/components/company-list-card/index.tsx
@@ -1,13 +1,7 @@
 import React, { useCallback } from "react";
 import { useNavigate } from "react-router-dom";
 
-import {
-	DefaultAvatar,
-	Icon,
-	IconButton,
-	StarRating,
-} from "~/common/components/index";
-import { IMAGE_UPLOAD_URL, LOGO_NONE } from "~/common/constants/index";
+import { Icon, IconButton, Logo, StarRating } from "~/common/components/index";
 import { AppRoute, IconName, RatingSize } from "~/common/enums/index";
 import { Company } from "~/common/types/index";
 
@@ -28,18 +22,11 @@ const CompanyListCard: React.FC<Properties> = ({ company }) => {
 		<div className={styles["company_list_card__container"]}>
 			<div className={styles["company_list_card__header"]}>
 				<div className={styles["company_list_card__logo_container"]}>
-					{company.logo !== LOGO_NONE ? (
-						<img
-							alt={`${company.name} logo`}
-							className={styles["company_list_card__logo"]}
-							src={`${IMAGE_UPLOAD_URL}${company.logo}`}
-						/>
-					) : (
-						<DefaultAvatar
-							className={styles["company_list_card__logo"]}
-							name={company.name}
-						/>
-					)}
+					<Logo
+						className={styles["company_list_card__logo"]}
+						logo={company.logo}
+						name={company.name}
+					/>
 					<span className={styles["company_list_card__name"]}>
 						{company.name}
 					</span>
