+++ src/pages/companies-list-page/components/companies-list/components/filtered-companies-list/components/company-card/index.tsx
@@ -2,12 +2,8 @@ import React from "react";
 import { Link } from "react-router-dom";
 
 import ShieldTick from "~/assets/images/shield-tick.svg?react";
-import { DefaultAvatar, StarRating } from "~/common/components/index";
-import {
-	IMAGE_UPLOAD_URL,
-	LOGO_NONE,
-	ScreenBreakpoints,
-} from "~/common/constants/index";
+import { Logo, StarRating } from "~/common/components/index";
+import { ScreenBreakpoints } from "~/common/constants/index";
 import { AppRoute, RatingSize } from "~/common/enums/index";
 import { useGetScreenWidth } from "~/common/hooks";
 import { Company } from "~/common/types/index";
@@ -28,18 +24,11 @@ const CompanyCard: React.FC<CompanyCardProperties> = ({ company }) => {
 			to={`${AppRoute.COMPANY_DETAILS}${company.id}`}
 		>
 			<div className={styles["company_card__logo_container"]}>
-				{company.logo !== LOGO_NONE ? (
-					<img
-						alt={`${company.name} logo`}
-						className={styles["company_card__company_logo"]}
-						src={`${IMAGE_UPLOAD_URL}${company.logo}`}
-					/>
-				) : (
-					<DefaultAvatar
-						className={styles["company_card__company_logo"]}
-						name={company.name}
-					/>
-				)}
+				<Logo
+					className={styles["company_card__company_logo"]}
+					logo={company.logo}
+					name={company.name}
+				/>
 				<div className={styles["company_card__rating_container"]}>
 					<StarRating
 						averageRating={company.avg_overall_rating}
