+++ src/pages/home-page/components/main-content/components/companies-section/components/company-card/index.tsx
@@ -1,8 +1,7 @@
 import React from "react";
 import { Link } from "react-router-dom";
 
-import { DefaultAvatar, StarRating } from "~/common/components/index";
-import { IMAGE_UPLOAD_URL, LOGO_NONE } from "~/common/constants/index";
+import { Logo, StarRating } from "~/common/components/index";
 import { AppRoute, RatingSize, StarRatingVariant } from "~/common/enums/index";
 import { Company } from "~/common/types/index";
 
@@ -30,19 +29,11 @@ const CompanyCard: React.FC<CompanyCardProperties> = ({ company }) => {
 				/>
 			</div>
 			<div className={styles["item_logo_container"]}>
-				{company.logo !== LOGO_NONE ? (
-					<img
-						alt={`${company.name} logo`}
-						className={styles["company_logo"]}
-						src={`${IMAGE_UPLOAD_URL}${company.logo}`}
-					/>
-				) : (
-					<DefaultAvatar
-						className={styles["company_logo"]}
-						name={company.name}
-					/>
-				)}
-
+				<Logo
+					className={styles["company_logo"]}
+					logo={company.logo}
+					name={company.name}
+				/>
 				<h4 className={styles["item_name"]}>{company.name}</h4>
 			</div>
 		</Link>
