+++ src/pages/home-page/components/main-content/components/course-card/index.tsx
@@ -2,8 +2,7 @@ import clsx from "clsx";
 import React from "react";
 import { Link } from "react-router-dom";
 
-import { DefaultAvatar, StarRating } from "~/common/components/index";
-import { IMAGE_UPLOAD_URL, LOGO_NONE } from "~/common/constants/index";
+import { Logo, StarRating } from "~/common/components/index";
 import { AppRoute, RatingSize, StarRatingVariant } from "~/common/enums/index";
 import { Course } from "~/common/types/index";
 
@@ -22,18 +21,11 @@ const CourseCard: React.FC<CourseCardProperties> = ({ className, course }) => {
 			to={`${AppRoute.COURSE_DETAILS}${course.id}`}
 		>
 			<div className={styles["item_loco_container"]}>
-				{course.company_logo !== LOGO_NONE ? (
-					<img
-						alt="$course.title} course - company logo"
-						className={styles["course_image"]}
-						src={`${IMAGE_UPLOAD_URL}${course.company_logo}`}
-					/>
-				) : (
-					<DefaultAvatar
-						className={styles["course_image"]}
-						name={course.company}
-					/>
-				)}
+				<Logo
+					className={styles["course_image"]}
+					logo={course.company_logo}
+					name={course.company}
+				/>
 				<div className={styles["item_rating_container"]}>
 					<StarRating
 						averageRating={course.avg_rating}
